Remove defaults for coredns_config
Currently dns server information need to be specified
as part of addon config and corresponding coredns_config.
However microk8s can handle updates in coredns_config if
addon specified the dns server information.

This patch removes the defaults of coredns_config
where the forward is hardcoded to 8.8.8.8,8.8.4.4 so
that microk8s can populate the correct configmap for
coredns.

Fixes: #66